tests/wpt/web-platform-tests/html/semantics/scripting-1/the-script-element/script-charset-01.html

<!DOCTYPE html>
<title>Script @type: unknown parameters</title>
<link rel="author" title="askalski" href="github.com/askalski">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#scriptingLanguages">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script type="text/javascript" id="external-script-windows1250-step1"
  src="serve-with-content-type.py?filename=external-script-windows1250.js&ct=text/javascript%3Bencoding=UTF8" charset="windows-1250">
</script>
<script>
//the charset is set correctly (step1)
test(function() {
  assert_equals(window.getSomeString(), "śćążź");
});
</script>
<script type="text/javascript" id="external-script-windows1250" src="external-script-windows1250.js" charset="windows-1250">
</script>
<script>
//the charset is set correctly (step1)
test(function() {
  assert_equals(window.getSomeString(), "śćążź");
});
</script>
<script type="text/javascript" id="external-script-windows1250-guess" src="external-script-windows1250.js">
</script>
<script>
//with no
test(function() {
  assert_not_equals(window.getSomeString(), "śćążź");
});
</script>

<!-- <script type="text/javascript" id="external-script-windows1252" src="external-script-windows1252.js" charset="windows-1252">
</script>
</div>
<script>
test(function() {
  assert_equals(document.getElementById("external-script-windows1251").charset, "windows-1251");
  assert_equals(document.getElementById("external-script-windows1252").charset, "windows-1252");
})
</script> -->
